Create loic
In the first part of my analysis I determine the rate of survivors  according to the  sex, class and place of embarkation. this allows me to find the sex, class and place of embarkation with the highest mortality rate, for example. in the second part I observe the behaviour of the ticket price   according to the state of the persons (survive or not survive), sex and class. I also observe the distribution of sex  according to the place of embarkation and the sex and the distribution of classes according to the place of embarkation.